pyproject.toml
[tool.poetry]
name = "bellsprout"
version = "0.1.0"
description = ""
authors = ["paul <paul@datalogue.io>"]

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "^4.6"
requests = "^2.20"
imageio = "^2.4"
numpy = "^1.15"
lxml = "^4.2"
pillow = "^5.3"
metar = "^1.6"
pyyaml = "^3.13"
python-arango = "^4.2"
pandas = "^0.23.4"
dask = "^0.20.2"
toolz = "^0.9.0"
cloudpickle = "^0.6.1"
phue = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
